+++ src/search/landmarks/landmark_factory.cc
@@ -2,7 +2,6 @@
 
 #include "landmark_graph.h"
 
-#include "exploration.h"
 #include "util.h"
 
 #include "../option_parser.h"
@@ -20,14 +19,16 @@ using namespace std;
 
 namespace landmarks {
 LandmarkFactory::LandmarkFactory(const options::Options &opts)
-    : lm_graph_task(nullptr),
-      reasonable_orders(opts.get<bool>("reasonable_orders")),
+    : reasonable_orders(opts.get<bool>("reasonable_orders")),
       only_causal_landmarks(opts.get<bool>("only_causal_landmarks")),
       disjunctive_landmarks(opts.get<bool>("disjunctive_landmarks")),
       conjunctive_landmarks(opts.get<bool>("conjunctive_landmarks")),
-      no_orders(opts.get<bool>("no_orders")) {
+      no_orders(opts.get<bool>("no_orders")),
+      lm_graph_task(nullptr) {
 }
 /*
+  TODO: Update this comment
+
   Note: To allow reusing landmark graphs, we use the following temporary
   solution.
 
@@ -60,65 +61,23 @@ shared_ptr<LandmarkGraph> LandmarkFactory::compute_lm_graph(
     lm_graph_task = task.get();
     utils::Timer lm_generation_timer;
 
+    lm_graph = make_shared<LandmarkGraph>();
+
     TaskProxy task_proxy(*task);
-
-    lm_graph = make_shared<LandmarkGraph>(task_proxy);
     generate_operators_lookups(task_proxy);
-    Exploration exploration(task_proxy);
-    generate_landmarks(task, exploration);
+    generate_landmarks(task);
 
-    // the following replaces the old "build_lm_graph"
-    generate(task_proxy, exploration);
     utils::g_log << "Landmarks generation time: " << lm_generation_timer << endl;
-    if (lm_graph->number_of_landmarks() == 0)
+    if (lm_graph->get_num_landmarks() == 0)
         utils::g_log << "Warning! No landmarks found. Task unsolvable?" << endl;
     else {
-        utils::g_log << "Discovered " << lm_graph->number_of_landmarks()
-                     << " landmarks, of which " << lm_graph->number_of_disj_landmarks()
+        utils::g_log << "Discovered " << lm_graph->get_num_landmarks()
+                     << " landmarks, of which " << lm_graph->get_num_disjunctive_landmarks()
                      << " are disjunctive and "
-                     << lm_graph->number_of_conj_landmarks() << " are conjunctive." << endl;
-        utils::g_log << lm_graph->number_of_edges() << " edges" << endl;
+                     << lm_graph->get_num_conjunctive_landmarks() << " are conjunctive." << endl;
+        utils::g_log << lm_graph->get_num_edges() << " edges" << endl;
     }
-    //lm_graph->dump();
     return lm_graph;
-}
-
-void LandmarkFactory::generate(const TaskProxy &task_proxy, Exploration &exploration) {
-    if (only_causal_landmarks)
-        discard_noncausal_landmarks(task_proxy, exploration);
-    if (!disjunctive_landmarks)
-        discard_disjunctive_landmarks();
-    if (!conjunctive_landmarks)
-        discard_conjunctive_landmarks();
-    lm_graph->set_landmark_ids();
-
-    if (no_orders)
-        discard_all_orderings();
-    else if (reasonable_orders) {
-        utils::g_log << "approx. reasonable orders" << endl;
-        approximate_reasonable_orders(task_proxy, false);
-        utils::g_log << "approx. obedient reasonable orders" << endl;
-        approximate_reasonable_orders(task_proxy, true);
-    }
-    mk_acyclic_graph();
-    calc_achievers(task_proxy, exploration);
-}
-
-bool LandmarkFactory::achieves_non_conditional(const OperatorProxy &o,
-                                               const LandmarkNode *lmp) const {
-    /* Test whether the landmark is achieved by the operator unconditionally.
-    A disjunctive landmarks is achieved if one of its disjuncts is achieved. */
-    assert(lmp);
-    for (EffectProxy effect: o.get_effects()) {
-        for (const FactPair &lm_fact : lmp->facts) {
-            FactProxy effect_fact = effect.get_fact();
-            if (effect_fact.get_pair() == lm_fact) {
-                if (effect.get_conditions().empty())
-                    return true;
-            }
-        }
-    }
-    return false;
 }
 
 bool LandmarkFactory::is_landmark_precondition(const OperatorProxy &op,
@@ -132,105 +91,6 @@ bool LandmarkFactory::is_landmark_precondition(const OperatorProxy &op,
                 return true;
         }
     }
-    return false;
-}
-
-bool LandmarkFactory::relaxed_task_solvable(const TaskProxy &task_proxy,
-                                            Exploration &exploration,
-                                            vector<vector<int>> &lvl_var,
-                                            vector<utils::HashMap<FactPair, int>> &lvl_op,
-                                            bool level_out, const LandmarkNode *exclude, bool compute_lvl_op) const {
-    /* Test whether the relaxed planning task is solvable without achieving the propositions in
-     "exclude" (do not apply operators that would add a proposition from "exclude").
-     As a side effect, collect in lvl_var and lvl_op the earliest possible point in time
-     when a proposition / operator can be achieved / become applicable in the relaxed task.
-     */
-
-    OperatorsProxy operators = task_proxy.get_operators();
-    AxiomsProxy axioms = task_proxy.get_axioms();
-    // Initialize lvl_op and lvl_var to numeric_limits<int>::max()
-    if (compute_lvl_op) {
-        lvl_op.resize(operators.size() + axioms.size());
-        for (OperatorProxy op : operators) {
-            add_operator_and_propositions_to_list(op, lvl_op);
-        }
-        for (OperatorProxy axiom : axioms) {
-            add_operator_and_propositions_to_list(axiom, lvl_op);
-        }
-    }
-    VariablesProxy variables = task_proxy.get_variables();
-    lvl_var.resize(variables.size());
-    for (VariableProxy var : variables) {
-        lvl_var[var.get_id()].resize(var.get_domain_size(),
-                                     numeric_limits<int>::max());
-    }
-    // Extract propositions from "exclude"
-    unordered_set<int> exclude_op_ids;
-    vector<FactPair> exclude_props;
-    if (exclude) {
-        for (OperatorProxy op : operators) {
-            if (achieves_non_conditional(op, exclude))
-                exclude_op_ids.insert(op.get_id());
-        }
-        exclude_props.insert(exclude_props.end(),
-                             exclude->facts.begin(), exclude->facts.end());
-    }
-    // Do relaxed exploration
-    exploration.compute_reachability_with_excludes(
-        lvl_var, lvl_op, level_out, exclude_props, exclude_op_ids, compute_lvl_op);
-
-    // Test whether all goal propositions have a level of less than numeric_limits<int>::max()
-    for (FactProxy goal : task_proxy.get_goals())
-        if (lvl_var[goal.get_variable().get_id()][goal.get_value()] ==
-            numeric_limits<int>::max())
-            return false;
-
-    return true;
-}
-
-void LandmarkFactory::add_operator_and_propositions_to_list(const OperatorProxy &op,
-                                                            vector<utils::HashMap<FactPair, int>> &lvl_op) const {
-    int op_or_axiom_id = get_operator_or_axiom_id(op);
-    for (EffectProxy effect : op.get_effects()) {
-        lvl_op[op_or_axiom_id].emplace(effect.get_fact().get_pair(), numeric_limits<int>::max());
-    }
-}
-
-bool LandmarkFactory::is_causal_landmark(const TaskProxy &task_proxy, Exploration &exploration,
-                                         const LandmarkNode &landmark) const {
-    /* Test whether the relaxed planning task is unsolvable without using any operator
-       that has "landmark" has a precondition.
-       Similar to "relaxed_task_solvable" above.
-     */
-
-    if (landmark.in_goal)
-        return true;
-    vector<vector<int>> lvl_var;
-    vector<utils::HashMap<FactPair, int>> lvl_op;
-    // Initialize lvl_var to numeric_limits<int>::max()
-    VariablesProxy variables = task_proxy.get_variables();
-    lvl_var.resize(variables.size());
-    for (VariableProxy var : variables) {
-        lvl_var[var.get_id()].resize(var.get_domain_size(),
-                                     numeric_limits<int>::max());
-    }
-    unordered_set<int> exclude_op_ids;
-    vector<FactPair> exclude_props;
-    for (OperatorProxy op : task_proxy.get_operators()) {
-        if (is_landmark_precondition(op, &landmark)) {
-            exclude_op_ids.insert(op.get_id());
-        }
-    }
-    // Do relaxed exploration
-    exploration.compute_reachability_with_excludes(
-        lvl_var, lvl_op, true, exclude_props, exclude_op_ids, false);
-
-    // Test whether all goal propositions have a level of less than numeric_limits<int>::max()
-    for (FactProxy goal : task_proxy.get_goals())
-        if (lvl_var[goal.get_variable().get_id()][goal.get_value()] ==
-            numeric_limits<int>::max())
-            return true;
-
     return false;
 }
 
@@ -482,7 +342,7 @@ void LandmarkFactory::approximate_reasonable_orders(
         if (obedient_orders && node_p->is_true_in_state(initial_state))
             continue;
 
-        if (!obedient_orders && node_p->is_goal()) {
+        if (!obedient_orders && node_p->is_true_in_goal) {
             for (auto &node2_p : lm_graph->get_nodes()) {
                 if (node2_p == node_p || node2_p->disjunctive)
                     continue;
@@ -491,7 +351,7 @@ void LandmarkFactory::approximate_reasonable_orders(
                     continue;
                 if (interferes(task_proxy, node2_p.get(), node_p.get())
                     && !have_common_achiever(node2_p.get(), node_p.get())) {
-                    edge_add(*node2_p, *node_p, EdgeType::reasonable);
+                    edge_add(*node2_p, *node_p, EdgeType::REASONABLE);
                 }
             }
         } else {
@@ -501,13 +361,13 @@ void LandmarkFactory::approximate_reasonable_orders(
             for (const auto &child : node_p->children) {
                 const LandmarkNode &node2 = *child.first;
                 const EdgeType &edge2 = child.second;
-                if (edge2 >= EdgeType::greedy_necessary) { // found node2: node_p ->_gn node2
+                if (edge2 >= EdgeType::GREEDY_NECESSARY) { // found node2: node_p ->_gn node2
                     for (const auto &p : node2.parents) {   // find parent
                         LandmarkNode &parent = *(p.first);
                         const EdgeType &edge = p.second;
                         if (parent.disjunctive)
                             continue;
-                        if ((edge >= EdgeType::natural || (obedient_orders && edge == EdgeType::reasonable)) &&
+                        if ((edge >= EdgeType::NATURAL || (obedient_orders && edge == EdgeType::REASONABLE)) &&
                             &parent != node_p.get()) {  // find predecessors or parent and collect in
                             // "interesting nodes"
                             interesting_nodes.insert(&parent);
@@ -528,9 +388,9 @@ void LandmarkFactory::approximate_reasonable_orders(
                 if (interferes(task_proxy, node, node_p.get())
                     && !have_common_achiever(node, node_p.get())) {
                     if (!obedient_orders)
-                        edge_add(*node, *node_p, EdgeType::reasonable);
+                        edge_add(*node, *node_p, EdgeType::REASONABLE);
                     else
-                        edge_add(*node, *node_p, EdgeType::obedient_reasonable);
+                        edge_add(*node, *node_p, EdgeType::OBEDIENT_REASONABLE);
                 }
             }
         }
@@ -561,7 +421,7 @@ void LandmarkFactory::collect_ancestors(
     for (const auto &p : node.parents) {
         LandmarkNode &parent = *(p.first);
         const EdgeType &edge = p.second;
-        if (edge >= EdgeType::natural || (use_reasonable && edge == EdgeType::reasonable))
+        if (edge >= EdgeType::NATURAL || (use_reasonable && edge == EdgeType::REASONABLE))
             if (closed_nodes.count(&parent) == 0) {
                 open_nodes.push_back(&parent);
                 closed_nodes.insert(&parent);
@@ -573,7 +433,7 @@ void LandmarkFactory::collect_ancestors(
         for (const auto &p : node2.parents) {
             LandmarkNode &parent = *(p.first);
             const EdgeType &edge = p.second;
-            if (edge >= EdgeType::natural || (use_reasonable && edge == EdgeType::reasonable)) {
+            if (edge >= EdgeType::NATURAL || (use_reasonable && edge == EdgeType::REASONABLE)) {
                 if (closed_nodes.count(&parent) == 0) {
                     open_nodes.push_back(&parent);
                     closed_nodes.insert(&parent);
@@ -591,10 +451,10 @@ void LandmarkFactory::edge_add(LandmarkNode &from, LandmarkNode &to,
     reduce cycles. If the edge is already present, the stronger edge type wins.
     */
     assert(&from != &to);
-    assert(from.parents.find(&to) == from.parents.end() || type <= EdgeType::reasonable);
-    assert(to.children.find(&from) == to.children.end() || type <= EdgeType::reasonable);
+    assert(from.parents.find(&to) == from.parents.end() || type <= EdgeType::REASONABLE);
+    assert(to.children.find(&from) == to.children.end() || type <= EdgeType::REASONABLE);
 
-    if (type == EdgeType::reasonable || type == EdgeType::obedient_reasonable) { // simple cycle test
+    if (type == EdgeType::REASONABLE || type == EdgeType::OBEDIENT_REASONABLE) { // simple cycle test
         if (from.parents.find(&to) != from.parents.end()) { // Edge in opposite direction exists
             //utils::g_log << "edge in opposite direction exists" << endl;
             if (from.parents.find(&to)->second > type) // Stronger order present, return
@@ -626,25 +486,14 @@ void LandmarkFactory::edge_add(LandmarkNode &from, LandmarkNode &to,
     assert(to.parents.find(&from) != to.parents.end());
 }
 
-void LandmarkFactory::discard_noncausal_landmarks(const TaskProxy &task_proxy, Exploration &exploration) {
-    int num_all_landmarks = lm_graph->number_of_landmarks();
-    lm_graph->remove_node_if(
-        [this, &task_proxy, &exploration](const LandmarkNode &node) {
-            return !is_causal_landmark(task_proxy, exploration, node);
-        });
-    int num_causal_landmarks = lm_graph->number_of_landmarks();
-    utils::g_log << "Discarded " << num_all_landmarks - num_causal_landmarks
-                 << " non-causal landmarks" << endl;
-}
-
 void LandmarkFactory::discard_disjunctive_landmarks() {
     /*
       Using disjunctive landmarks during landmark generation can be beneficial
       even if we don't want to use disjunctive landmarks during search. So we
       allow removing disjunctive landmarks after landmark generation.
     */
-    if (lm_graph->number_of_disj_landmarks() > 0) {
-        utils::g_log << "Discarding " << lm_graph->number_of_disj_landmarks()
+    if (lm_graph->get_num_disjunctive_landmarks() > 0) {
+        utils::g_log << "Discarding " << lm_graph->get_num_disjunctive_landmarks()
                      << " disjunctive landmarks" << endl;
         lm_graph->remove_node_if(
             [](const LandmarkNode &node) {return node.disjunctive;});
@@ -652,8 +501,8 @@ void LandmarkFactory::discard_disjunctive_landmarks() {
 }
 
 void LandmarkFactory::discard_conjunctive_landmarks() {
-    if (lm_graph->number_of_conj_landmarks() > 0) {
-        utils::g_log << "Discarding " << lm_graph->number_of_conj_landmarks()
+    if (lm_graph->get_num_conjunctive_landmarks() > 0) {
+        utils::g_log << "Discarding " << lm_graph->get_num_conjunctive_landmarks()
                      << " conjunctive landmarks" << endl;
         lm_graph->remove_node_if(
             [](const LandmarkNode &node) {return node.conjunctive;});
@@ -669,7 +518,7 @@ void LandmarkFactory::discard_all_orderings() {
 }
 
 void LandmarkFactory::mk_acyclic_graph() {
-    unordered_set<LandmarkNode *> acyclic_node_set(lm_graph->number_of_landmarks());
+    unordered_set<LandmarkNode *> acyclic_node_set(lm_graph->get_num_landmarks());
     int removed_edges = 0;
     for (auto &node : lm_graph->get_nodes()) {
         if (acyclic_node_set.find(node.get()) == acyclic_node_set.end())
@@ -690,7 +539,7 @@ bool LandmarkFactory::remove_first_weakest_cycle_edge(LandmarkNode *cur,
     for (list<pair<LandmarkNode *, EdgeType>>::iterator it2 = it; it2
          != path.end(); ++it2) {
         EdgeType edge = it2->second;
-        if (edge == EdgeType::reasonable || edge == EdgeType::obedient_reasonable) {
+        if (edge == EdgeType::REASONABLE || edge == EdgeType::OBEDIENT_REASONABLE) {
             parent_p = it2->first;
             if (*it2 == path.back()) {
                 child_p = cur;
@@ -700,7 +549,7 @@ bool LandmarkFactory::remove_first_weakest_cycle_edge(LandmarkNode *cur,
                 ++child_it;
                 child_p = child_it->first;
             }
-            if (edge == EdgeType::obedient_reasonable)
+            if (edge == EdgeType::OBEDIENT_REASONABLE)
                 break;
             // else no break since o_r order could still appear in list
         }
@@ -718,7 +567,7 @@ int LandmarkFactory::loop_acyclic_graph(LandmarkNode &lmn,
     assert(acyclic_node_set.find(&lmn) == acyclic_node_set.end());
     int nr_removed = 0;
     list<pair<LandmarkNode *, EdgeType>> path;
-    unordered_set<LandmarkNode *> visited = unordered_set<LandmarkNode *>(lm_graph->number_of_landmarks());
+    unordered_set<LandmarkNode *> visited = unordered_set<LandmarkNode *>(lm_graph->get_num_landmarks());
     LandmarkNode *cur = &lmn;
     while (true) {
         assert(acyclic_node_set.find(cur) == acyclic_node_set.end());
@@ -769,44 +618,6 @@ int LandmarkFactory::loop_acyclic_graph(LandmarkNode &lmn,
     return nr_removed;
 }
 
-void LandmarkFactory::compute_predecessor_information(
-    const TaskProxy &task_proxy,
-    Exploration &exploration,
-    LandmarkNode *bp,
-    vector<vector<int>> &lvl_var,
-    vector<utils::HashMap<FactPair, int>> &lvl_op) {
-    /* Collect information at what time step propositions can be reached
-    (in lvl_var) in a relaxed plan that excludes bp, and similarly
-    when operators can be applied (in lvl_op).  */
-
-    relaxed_task_solvable(task_proxy, exploration, lvl_var, lvl_op, true, bp);
-}
-
-void LandmarkFactory::calc_achievers(const TaskProxy &task_proxy, Exploration &exploration) {
-    VariablesProxy variables = task_proxy.get_variables();
-    for (auto &lmn : lm_graph->get_nodes()) {
-        for (const FactPair &lm_fact : lmn->facts) {
-            const vector<int> &ops = get_operators_including_eff(lm_fact);
-            lmn->possible_achievers.insert(ops.begin(), ops.end());
-
-            if (variables[lm_fact.var].is_derived())
-                lmn->is_derived = true;
-        }
-
-        vector<vector<int>> lvl_var;
-        vector<utils::HashMap<FactPair, int>> lvl_op;
-        compute_predecessor_information(task_proxy, exploration, lmn.get(), lvl_var, lvl_op);
-
-        for (int op_or_axom_id : lmn->possible_achievers) {
-            OperatorProxy op = get_operator_or_axiom(task_proxy, op_or_axom_id);
-
-            if (_possibly_reaches_lm(op, lvl_var, lmn.get())) {
-                lmn->first_achievers.insert(op_or_axom_id);
-            }
-        }
-    }
-}
-
 void LandmarkFactory::generate_operators_lookups(const TaskProxy &task_proxy) {
     /* Build datastructures for efficient landmark computation. Map propositions
     to the operators that achieve them or have them as preconditions */
@@ -834,7 +645,6 @@ void LandmarkFactory::generate_operators_lookups(const TaskProxy &task_proxy) {
         }
     }
 }
-
 
 void _add_options_to_parser(OptionParser &parser) {
     parser.add_option<bool>("reasonable_orders",