+++ src/search/landmarks/landmark_status_manager.cc
@@ -12,14 +12,14 @@ namespace landmarks {
   computing new landmark information.
 */
 LandmarkStatusManager::LandmarkStatusManager(LandmarkGraph &graph)
-    : reached_lms(vector<bool>(graph.number_of_landmarks(), true)),
-      lm_status(graph.number_of_landmarks(), lm_not_reached),
+    : reached_lms(vector<bool>(graph.get_num_landmarks(), true)),
+      lm_status(graph.get_num_landmarks(), lm_not_reached),
       lm_graph(graph) {
 }
 
 landmark_status LandmarkStatusManager::get_landmark_status(
     size_t id) const {
-    assert(static_cast<int>(id) < lm_graph.number_of_landmarks());
+    assert(static_cast<int>(id) < lm_graph.get_num_landmarks());
     return lm_status[id];
 }
 
@@ -36,7 +36,7 @@ void LandmarkStatusManager::set_landmarks_for_initial_state(
     int inserted = 0;
     int num_goal_lms = 0;
     for (auto &node_p : lm_graph.get_nodes()) {
-        if (node_p->in_goal) {
+        if (node_p->is_true_in_goal) {
             ++num_goal_lms;
         }
 
@@ -81,7 +81,7 @@ bool LandmarkStatusManager::update_reached_lms(
         parent_ancestor_state);
     BitsetView reached = get_reached_landmarks(ancestor_state);
 
-    int num_landmarks = lm_graph.number_of_landmarks();
+    int num_landmarks = lm_graph.get_num_landmarks();
     assert(reached.size() == num_landmarks);
     assert(parent_reached.size() == num_landmarks);
 
@@ -101,7 +101,7 @@ bool LandmarkStatusManager::update_reached_lms(
     // Mark landmarks reached right now as "reached" (if they are "leaves").
     for (int id = 0; id < num_landmarks; ++id) {
         if (!reached.test(id)) {
-            LandmarkNode *node = lm_graph.get_lm_for_index(id);
+            LandmarkNode *node = lm_graph.get_landmark(id);
             if (node->is_true_in_state(ancestor_state)) {
                 reached.set(id);
             }
@@ -116,10 +116,10 @@ void LandmarkStatusManager::update_lm_status(const State &ancestor_state) {
 
     /* This first loop is necessary as setup for the *needed again*
        check in the second loop. */
-    for (int id = 0; id < lm_graph.number_of_landmarks(); ++id) {
+    for (int id = 0; id < lm_graph.get_num_landmarks(); ++id) {
         lm_status[id] = reached.test(id) ? lm_reached : lm_not_reached;
     }
-    for (int id = 0; id < lm_graph.number_of_landmarks(); ++id) {
+    for (int id = 0; id < lm_graph.get_num_landmarks(); ++id) {
         if (lm_status[id] == lm_reached
             && landmark_needed_again(id, ancestor_state)) {
             lm_status[id] = lm_needed_again;
@@ -158,10 +158,10 @@ bool LandmarkStatusManager::dead_end_exists() {
 
 bool LandmarkStatusManager::landmark_needed_again(
     int id, const State &state) {
-    LandmarkNode *node = lm_graph.get_lm_for_index(id);
+    LandmarkNode *node = lm_graph.get_landmark(id);
     if (node->is_true_in_state(state)) {
         return has_unreached_parent(node);
-    } else if (node->is_goal()) {
+    } else if (node->is_true_in_goal) {
         return true;
     } else {
         /*
@@ -170,7 +170,7 @@ bool LandmarkStatusManager::landmark_needed_again(
           achieving B for the first time, it must become true again.
         */
         for (const auto &child : node->children) {
-            if (child.second >= EdgeType::greedy_necessary
+            if (child.second >= EdgeType::GREEDY_NECESSARY
                 && lm_status[child.first->get_id()] == lm_not_reached) {
                 return true;
             }
@@ -184,7 +184,7 @@ bool LandmarkStatusManager::has_unreached_parent(
     for (const auto &parent : node->parents) {
         if (lm_status[parent.first->get_id()] == lm_not_reached) {
             // This cannot occur for natural orderings by definition.
-            assert(parent.second < EdgeType::natural);
+            assert(parent.second < EdgeType::NATURAL);
             return true;
         }
     }